server/requirements.txt
Django>=2.0,<3.0
django-extensions
psycopg2-binary>=2.8
django-admin-sortable2
djangorestframework
django-cors-headers